fix(publish): stage list add publish validate
